+++ NCCRD.Clients.Web.React/app/js/components/Base/CustomNavbar.jsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { BeatLoader } from 'react-spinners'
 import { Button, Input } from 'mdbreact'
-import * as ACTION_TYPES from "../constants/action-types"
+import * as ACTION_TYPES from "../../constants/action-types"
 import { Navbar, NavbarBrand, NavbarNav, NavbarToggler, Collapse, NavItem, NavLink } from 'mdbreact';
 
 const mapStateToProps = (state, props) => {
