+++ NCCRD.Clients.Web.React/app/js/components/Base/Home.jsx
