+++ NCCRD.Clients.Web.React/app/js/components/Projects/Details/AdaptationDetailsItem.jsx
@@ -1,11 +1,11 @@
 'use strict'
 
 import React from 'react'
-import { apiBaseURL } from "../constants/apiBaseURL"
+import { apiBaseURL } from "../../../constants/apiBaseURL"
 import { connect } from 'react-redux'
-import TextAreaComponent from './TextAreaComponent.jsx'
-import SelectComponent from './SelectComponent.jsx'
-import * as ACTION_TYPES from "../constants/action-types"
+import TextAreaComponent from '../../Shared/TextAreaComponent.jsx'
+import SelectComponent from '../../Shared/SelectComponent.jsx'
+import * as ACTION_TYPES from "../../../constants/action-types"
 
 const mapStateToProps = (state, props) => {
   let { lookupData: { adaptationPurpose, sector, sectorTree, sectorType, typology } } = state
