+++ NCCRD.Clients.Web.React/app/js/components/Projects/Details/ResearchDetailsTab.jsx
@@ -2,9 +2,9 @@
 
 import React from 'react'
 import { Button } from 'mdbreact'
-import { apiBaseURL } from "../constants/apiBaseURL"
+import { apiBaseURL } from "../../../constants/apiBaseURL"
 import { connect } from 'react-redux'
-import * as ACTION_TYPES from "../constants/action-types"
+import * as ACTION_TYPES from "../../../constants/action-types"
 import ResearchDetailsItem from './ResearchDetailsItem.jsx'
 
 const mapStateToProps = (state, props) => {
