+++ NCCRD.Clients.Web.React/app/js/components/Projects/List/ProjectCard.jsx
