+++ NCCRD.Clients.Web.React/app/js/components/Shared/RangeComponent.jsx
