+++ NCCRD.Clients.Web.React/app/js/components/Shared/TextComponent.jsx
