+++ NCCRD.Clients.Web.React/app/js/index.jsx
@@ -15,12 +15,12 @@ import store from './store'
 import queryString from 'query-string'
 import { Button } from 'mdbreact/'
 import { HashRouter as Router, Switch, Route, Redirect } from 'react-router-dom'
-import Home from './components/Home.jsx'
-import Projects from './components/Projects.jsx'
-import ProjectDetails from './components/ProjectDetails.jsx'
-import Login from './components/Login.jsx'
-import Logout from './components/Logout.jsx'
-import CustomNavbar from './components/CustomNavbar.jsx'
+import Home from './components/Base/Home.jsx'
+import Projects from './components/Projects/List/Projects.jsx'
+import ProjectDetails from './components/Projects/Details/ProjectDetails.jsx'
+import Login from './components/Authentication/Login.jsx'
+import Logout from './components/Authentication/Logout.jsx'
+import CustomNavbar from './components/Base/CustomNavbar.jsx'
 
 /**
  * Tap Event
